Fix bug where LibraryActivity was finished when opening document.
This meant that the LibraryActivity could not be returned to
to open a second document. There is no reason to finish the
LibraryActivity when opening a document since we'll either
return to it when the user presses Back and the history is empty,
or the user may manually switch to that activity themselves.