Show error message when outline depth/length is truncated.
When opening pdfref10.pdf the app crashes due to an exception. This is
because the outline data in the Bundle sent from DocumentActivity to
OutlineActivity is larger than 1Mbyte, which is Android's arbitrary
Bundle size limit.

The exception cannot be caught in OutlineActivity becuase it is not
yet started. The exception surprisngly can't be caught in
DocumentActivity that is starting the OutlineActivity either, because
the exception backtrace does not include any function in
DocumentActivity.

Instead the framework that passes Bundles between activities throws
the exception in the apps process causing the crash.

This commit just tries to avoid the crash by creating an arbitrary
outline count limit and outline depth limit.

An alternative would be to measure the Bundle size limit when each
outline item is inserted, but that means repeatedly updating the
Bundle contents querying its size for every added item, something
that is likely inefficient.

It may be possible to pass portions of outline items in multiple
Bundles, but how to do that is currently unknown.