Set maximum page label size based on maximum page count.
Previously the page scrubber had a fixed width of 110dp. This is problematic
when the number of pages is large because the entire string, currentPage + " /
" + pageCount, is not visible.

Here we instead allow the page label to wrap its content, but how wide can the
page label be? The largest width is when when last page of a document is shown.
By setting the minimum size of the page label to this width the page label will
stay the same size regardless of what page is being shown, which in turn fixes
the size of the page scrubber too.

So, measure the width of the text, pageCount + " / " + pageCount, and
using the same text rendering as the existing page label, add any
padding and set that as the minimum width. Do this when the page number is
updated anyway and remember the width.